feat: component directory
